+++ FirebaseVertexAI/Sample/ChatSample/Views/ErrorDetailsView.swift
@@ -16,28 +16,29 @@ import FirebaseVertexAI
 import MarkdownUI
 import SwiftUI
 
-extension SafetySetting.HarmCategory: CustomStringConvertible {
-  public var description: String {
+private extension HarmCategory {
+  /// Returns a description of the `HarmCategory` suitable for displaying in the UI.
+  var displayValue: String {
     switch self {
     case .dangerousContent: "Dangerous content"
     case .harassment: "Harassment"
     case .hateSpeech: "Hate speech"
     case .sexuallyExplicit: "Sexually explicit"
-    case .unknown: "Unknown"
-    case .unspecified: "Unspecified"
+    case .civicIntegrity: "Civic integrity"
+    default: "Unknown HarmCategory: \(rawValue)"
     }
   }
 }
 
-extension SafetyRating.HarmProbability: CustomStringConvertible {
-  public var description: String {
+private extension SafetyRating.HarmProbability {
+  /// Returns a description of the `HarmProbability` suitable for displaying in the UI.
+  var displayValue: String {
     switch self {
     case .high: "High"
     case .low: "Low"
     case .medium: "Medium"
     case .negligible: "Negligible"
-    case .unknown: "Unknown"
-    case .unspecified: "Unspecified"
+    default: "Unknown HarmProbability: \(rawValue)"
     }
   }
 }
@@ -75,10 +76,9 @@ private struct SafetyRatingsSection: View {
     Section("Safety ratings") {
       List(ratings, id: \.self) { rating in
         HStack {
-          Text("\(String(describing: rating.category))")
-            .font(.subheadline)
+          Text(rating.category.displayValue).font(.subheadline)
           Spacer()
-          Text("\(String(describing: rating.probability))")
+          Text(rating.probability.displayValue)
         }
       }
     }
@@ -162,12 +162,11 @@ struct ErrorDetailsView: View {
   let error = GenerateContentError.responseStoppedEarly(
     reason: .maxTokens,
     response: GenerateContentResponse(candidates: [
-      CandidateResponse(content: ModelContent(role: "model", [
+      CandidateResponse(content: ModelContent(role: "model", parts:
         """
         A _hypothetical_ model response.
         Cillum ex aliqua amet aliquip labore amet eiusmod consectetur reprehenderit sit commodo.
-        """,
-      ]),
+        """),
       safetyRatings: [
         SafetyRating(category: .dangerousContent, probability: .high),
         SafetyRating(category: .harassment, probability: .low),
@@ -185,12 +184,11 @@ struct ErrorDetailsView: View {
 #Preview("Prompt Blocked") {
   let error = GenerateContentError.promptBlocked(
     response: GenerateContentResponse(candidates: [
-      CandidateResponse(content: ModelContent(role: "model", [
+      CandidateResponse(content: ModelContent(role: "model", parts:
         """
         A _hypothetical_ model response.
         Cillum ex aliqua amet aliquip labore amet eiusmod consectetur reprehenderit sit commodo.
-        """,
-      ]),
+        """),
       safetyRatings: [
         SafetyRating(category: .dangerousContent, probability: .high),
         SafetyRating(category: .harassment, probability: .low),