+++ FirebaseVertexAI/Sources/FunctionCalling.swift
@@ -14,15 +14,6 @@
 
 import Foundation
 
-/// A predicted function call returned from the model.
-public struct FunctionCall: Equatable, Sendable {
-  /// The name of the function to call.
-  public let name: String
-
-  /// The function parameters and values.
-  public let args: JSONObject
-}
-
 /// Structured representation of a function declaration.
 ///
 /// This `FunctionDeclaration` is a representation of a block of code that can be used as a ``Tool``
@@ -43,17 +34,15 @@ public struct FunctionDeclaration {
   ///   - name: The name of the function; must be a-z, A-Z, 0-9, or contain underscores and dashes,
   ///   with a maximum length of 63.
   ///   - description: A brief description of the function.
-  ///   - parameters: Describes the parameters to this function; the keys are parameter names and
-  ///   the values are ``Schema`` objects describing them.
-  ///   - requiredParameters: A list of required parameters by name.
-  public init(name: String, description: String, parameters: [String: Schema]?,
-              requiredParameters: [String]? = nil) {
+  ///   - parameters: Describes the parameters to this function.
+  public init(name: String, description: String, parameters: [String: Schema],
+              optionalParameters: [String] = []) {
     self.name = name
     self.description = description
-    self.parameters = Schema(
-      type: .object,
+    self.parameters = Schema.object(
       properties: parameters,
-      requiredProperties: requiredParameters
+      optionalProperties: optionalParameters,
+      nullable: false
     )
   }
 }
@@ -126,49 +115,7 @@ public struct ToolConfig {
   }
 }
 
-/// Result output from a ``FunctionCall``.
-///
-/// Contains a string representing the `FunctionDeclaration.name` and a structured JSON object
-/// containing any output from the function is used as context to the model. This should contain the
-/// result of a ``FunctionCall`` made based on model prediction.
-public struct FunctionResponse: Equatable, Sendable {
-  /// The name of the function that was called.
-  let name: String
-
-  /// The function's response.
-  let response: JSONObject
-
-  /// Constructs a new `FunctionResponse`.
-  ///
-  /// - Parameters:
-  ///   - name: The name of the function that was called.
-  ///   - response: The function's response.
-  public init(name: String, response: JSONObject) {
-    self.name = name
-    self.response = response
-  }
-}
-
 // MARK: - Codable Conformance
-
-extension FunctionCall: Decodable {
-  enum CodingKeys: CodingKey {
-    case name
-    case args
-  }
-
-  public init(from decoder: Decoder) throws {
-    let container = try decoder.container(keyedBy: CodingKeys.self)
-    name = try container.decode(String.self, forKey: .name)
-    if let args = try container.decodeIfPresent(JSONObject.self, forKey: .args) {
-      self.args = args
-    } else {
-      args = JSONObject()
-    }
-  }
-}
-
-extension FunctionCall: Encodable {}
 
 extension FunctionDeclaration: Encodable {
   enum CodingKeys: String, CodingKey {
@@ -192,5 +139,3 @@ extension FunctionCallingConfig: Encodable {}
 extension FunctionCallingConfig.Mode: Encodable {}
 
 extension ToolConfig: Encodable {}
-
-extension FunctionResponse: Encodable {}