+++ FirebaseVertexAI/Sources/GenerativeAIRequest.swift
@@ -28,17 +28,18 @@ protocol GenerativeAIRequest: Encodable {
 public struct RequestOptions {
   /// The request’s timeout interval in seconds; if not specified uses the default value for a
   /// `URLRequest`.
-  let timeout: TimeInterval?
+  let timeout: TimeInterval
 
   /// The API version to use in requests to the backend.
+  ///
+  /// TODO(andrewheard): Update to "v1beta" after the Vertex AI in Firebase API launch.
   let apiVersion = "v2beta"
 
   /// Initializes a request options object.
   ///
   /// - Parameters:
-  ///   - timeout The request’s timeout interval in seconds; if not specified uses the default value
-  ///   for a `URLRequest`.
-  public init(timeout: TimeInterval? = nil) {
+  ///   - timeout The request’s timeout interval in seconds; defaults to 180 seconds.
+  public init(timeout: TimeInterval = 180.0) {
     self.timeout = timeout
   }
 }