+++ FirebaseVertexAI/Sources/Safety.swift
@@ -19,59 +19,69 @@ import Foundation
 /// responses that exceed a certain threshold.
 @available(iOS 15.0, macOS 11.0, macCatalyst 15.0, tvOS 15.0, watchOS 8.0, *)
 public struct SafetyRating: Equatable, Hashable, Sendable {
-  /// The category describing the potential harm a piece of content may pose. See
-  /// ``SafetySetting/HarmCategory`` for a list of possible values.
-  public let category: SafetySetting.HarmCategory
+  /// The category describing the potential harm a piece of content may pose.
+  ///
+  /// See ``HarmCategory`` for a list of possible values.
+  public let category: HarmCategory
 
-  /// The model-generated probability that a given piece of content falls under the harm category
-  /// described in ``SafetySetting/HarmCategory``. This does not indicate the severity of harm for a
-  /// piece of content. See ``HarmProbability`` for a list of possible values.
+  /// The model-generated probability that the content falls under the specified harm ``category``.
+  ///
+  /// See ``HarmProbability`` for a list of possible values.
+  ///
+  /// > Important: This does not indicate the severity of harm for a piece of content.
   public let probability: HarmProbability
 
   /// Initializes a new `SafetyRating` instance with the given category and probability.
   /// Use this initializer for SwiftUI previews or tests.
-  public init(category: SafetySetting.HarmCategory, probability: HarmProbability) {
+  public init(category: HarmCategory, probability: HarmProbability) {
     self.category = category
     self.probability = probability
   }
 
-  /// The probability that a given model output falls under a harmful content category. This does
-  /// not indicate the severity of harm for a piece of content.
-  public enum HarmProbability: String, Sendable {
-    /// Unknown. A new server value that isn't recognized by the SDK.
-    case unknown = "UNKNOWN"
+  /// The probability that a given model output falls under a harmful content category.
+  ///
+  /// > Note: This does not indicate the severity of harm for a piece of content.
+  public struct HarmProbability: DecodableProtoEnum, Hashable, Sendable {
+    enum Kind: String {
+      case negligible = "NEGLIGIBLE"
+      case low = "LOW"
+      case medium = "MEDIUM"
+      case high = "HIGH"
+    }
 
-    /// The probability was not specified in the server response.
-    case unspecified = "HARM_PROBABILITY_UNSPECIFIED"
-
-    /// The probability is zero or close to zero. For benign content, the probability across all
-    /// categories will be this value.
-    case negligible = "NEGLIGIBLE"
+    /// The probability is zero or close to zero.
+    ///
+    /// For benign content, the probability across all categories will be this value.
+    public static var negligible: HarmProbability {
+      return self.init(kind: .negligible)
+    }
 
     /// The probability is small but non-zero.
-    case low = "LOW"
+    public static var low: HarmProbability {
+      return self.init(kind: .low)
+    }
 
     /// The probability is moderate.
-    case medium = "MEDIUM"
+    public static var medium: HarmProbability {
+      return self.init(kind: .medium)
+    }
 
-    /// The probability is high. The content described is very likely harmful.
-    case high = "HIGH"
-  }
-}
+    /// The probability is high.
+    ///
+    /// The content described is very likely harmful.
+    public static var high: HarmProbability {
+      return self.init(kind: .high)
+    }
 
-/// Safety feedback for an entire request.
-@available(iOS 15.0, macOS 11.0, macCatalyst 15.0, tvOS 15.0, watchOS 8.0, *)
-public struct SafetyFeedback {
-  /// Safety rating evaluated from content.
-  public let rating: SafetyRating
+    /// Returns the raw string representation of the `HarmProbability` value.
+    ///
+    /// > Note: This value directly corresponds to the values in the [REST
+    /// > API](https://cloud.google.com/vertex-ai/docs/reference/rest/v1beta1/GenerateContentResponse#SafetyRating).
+    public let rawValue: String
 
-  /// Safety settings applied to the request.
-  public let setting: SafetySetting
-
-  /// Internal initializer.
-  init(rating: SafetyRating, setting: SafetySetting) {
-    self.rating = rating
-    self.setting = setting
+    var unrecognizedValueMessageCode: VertexLog.MessageCode {
+      .generateContentResponseUnrecognizedHarmProbability
+    }
   }
 }
 
@@ -79,47 +89,42 @@ public struct SafetyFeedback {
 /// fallback response instead of generated content.
 @available(iOS 15.0, macOS 11.0, macCatalyst 15.0, tvOS 15.0, watchOS 8.0, *)
 public struct SafetySetting {
-  /// A type describing safety attributes, which include harmful categories and topics that can
-  /// be considered sensitive.
-  public enum HarmCategory: String, Sendable {
-    /// Unknown. A new server value that isn't recognized by the SDK.
-    case unknown = "HARM_CATEGORY_UNKNOWN"
-
-    /// Unspecified by the server.
-    case unspecified = "HARM_CATEGORY_UNSPECIFIED"
-
-    /// Harassment content.
-    case harassment = "HARM_CATEGORY_HARASSMENT"
-
-    /// Negative or harmful comments targeting identity and/or protected attributes.
-    case hateSpeech = "HARM_CATEGORY_HATE_SPEECH"
-
-    /// Contains references to sexual acts or other lewd content.
-    case sexuallyExplicit = "HARM_CATEGORY_SEXUALLY_EXPLICIT"
-
-    /// Promotes or enables access to harmful goods, services, or activities.
-    case dangerousContent = "HARM_CATEGORY_DANGEROUS_CONTENT"
-  }
-
   /// Block at and beyond a specified ``SafetyRating/HarmProbability``.
-  public enum BlockThreshold: String, Sendable {
-    /// Unknown. A new server value that isn't recognized by the SDK.
-    case unknown = "UNKNOWN"
+  public struct HarmBlockThreshold: EncodableProtoEnum, Sendable {
+    enum Kind: String {
+      case blockLowAndAbove = "BLOCK_LOW_AND_ABOVE"
+      case blockMediumAndAbove = "BLOCK_MEDIUM_AND_ABOVE"
+      case blockOnlyHigh = "BLOCK_ONLY_HIGH"
+      case blockNone = "BLOCK_NONE"
+      case off = "OFF"
+    }
 
-    /// Threshold is unspecified.
-    case unspecified = "HARM_BLOCK_THRESHOLD_UNSPECIFIED"
-
-    // Content with `.negligible` will be allowed.
-    case blockLowAndAbove = "BLOCK_LOW_AND_ABOVE"
+    /// Content with `.negligible` will be allowed.
+    public static var blockLowAndAbove: HarmBlockThreshold {
+      return self.init(kind: .blockLowAndAbove)
+    }
 
     /// Content with `.negligible` and `.low` will be allowed.
-    case blockMediumAndAbove = "BLOCK_MEDIUM_AND_ABOVE"
+    public static var blockMediumAndAbove: HarmBlockThreshold {
+      return self.init(kind: .blockMediumAndAbove)
+    }
 
     /// Content with `.negligible`, `.low`, and `.medium` will be allowed.
-    case blockOnlyHigh = "BLOCK_ONLY_HIGH"
+    public static var blockOnlyHigh: HarmBlockThreshold {
+      return self.init(kind: .blockOnlyHigh)
+    }
 
     /// All content will be allowed.
-    case blockNone = "BLOCK_NONE"
+    public static var blockNone: HarmBlockThreshold {
+      return self.init(kind: .blockNone)
+    }
+
+    /// Turn off the safety filter.
+    public static var off: HarmBlockThreshold {
+      return self.init(kind: .off)
+    }
+
+    let rawValue: String
   }
 
   enum CodingKeys: String, CodingKey {
@@ -131,67 +136,68 @@ public struct SafetySetting {
   public let harmCategory: HarmCategory
 
   /// The threshold describing what content should be blocked.
-  public let threshold: BlockThreshold
+  public let threshold: HarmBlockThreshold
 
   /// Initializes a new safety setting with the given category and threshold.
-  public init(harmCategory: HarmCategory, threshold: BlockThreshold) {
+  public init(harmCategory: HarmCategory, threshold: HarmBlockThreshold) {
     self.harmCategory = harmCategory
     self.threshold = threshold
+  }
+}
+
+/// Categories describing the potential harm a piece of content may pose.
+public struct HarmCategory: CodableProtoEnum, Hashable, Sendable {
+  enum Kind: String {
+    case harassment = "HARM_CATEGORY_HARASSMENT"
+    case hateSpeech = "HARM_CATEGORY_HATE_SPEECH"
+    case sexuallyExplicit = "HARM_CATEGORY_SEXUALLY_EXPLICIT"
+    case dangerousContent = "HARM_CATEGORY_DANGEROUS_CONTENT"
+    case civicIntegrity = "HARM_CATEGORY_CIVIC_INTEGRITY"
+  }
+
+  /// Harassment content.
+  public static var harassment: HarmCategory {
+    return self.init(kind: .harassment)
+  }
+
+  /// Negative or harmful comments targeting identity and/or protected attributes.
+  public static var hateSpeech: HarmCategory {
+    return self.init(kind: .hateSpeech)
+  }
+
+  /// Contains references to sexual acts or other lewd content.
+  public static var sexuallyExplicit: HarmCategory {
+    return self.init(kind: .sexuallyExplicit)
+  }
+
+  /// Promotes or enables access to harmful goods, services, or activities.
+  public static var dangerousContent: HarmCategory {
+    return self.init(kind: .dangerousContent)
+  }
+
+  /// Content that may be used to harm civic integrity.
+  public static var civicIntegrity: HarmCategory {
+    return self.init(kind: .civicIntegrity)
+  }
+
+  /// Returns the raw string representation of the `HarmCategory` value.
+  ///
+  /// > Note: This value directly corresponds to the values in the
+  /// > [REST API](https://cloud.google.com/vertex-ai/docs/reference/rest/v1beta1/HarmCategory).
+  public let rawValue: String
+
+  var unrecognizedValueMessageCode: VertexLog.MessageCode {
+    .generateContentResponseUnrecognizedHarmCategory
   }
 }
 
 // MARK: - Codable Conformances
 
 @available(iOS 15.0, macOS 11.0, macCatalyst 15.0, tvOS 15.0, watchOS 8.0, *)
-extension SafetyRating.HarmProbability: Codable {
-  public init(from decoder: Decoder) throws {
-    let value = try decoder.singleValueContainer().decode(String.self)
-    guard let decodedProbability = SafetyRating.HarmProbability(rawValue: value) else {
-      Logging.default
-        .error("[FirebaseVertexAI] Unrecognized HarmProbability with value \"\(value)\".")
-      self = .unknown
-      return
-    }
-
-    self = decodedProbability
-  }
-}
-
-@available(iOS 15.0, macOS 11.0, macCatalyst 15.0, tvOS 15.0, watchOS 8.0, *)
 extension SafetyRating: Decodable {}
 
 @available(iOS 15.0, macOS 11.0, macCatalyst 15.0, tvOS 15.0, watchOS 8.0, *)
-extension SafetyFeedback: Decodable {}
+extension SafetySetting.HarmBlockThreshold: Encodable {}
 
 @available(iOS 15.0, macOS 11.0, macCatalyst 15.0, tvOS 15.0, watchOS 8.0, *)
-extension SafetySetting.HarmCategory: Codable {
-  public init(from decoder: Decoder) throws {
-    let value = try decoder.singleValueContainer().decode(String.self)
-    guard let decodedCategory = SafetySetting.HarmCategory(rawValue: value) else {
-      Logging.default
-        .error("[FirebaseVertexAI] Unrecognized HarmCategory with value \"\(value)\".")
-      self = .unknown
-      return
-    }
-
-    self = decodedCategory
-  }
-}
-
-@available(iOS 15.0, macOS 11.0, macCatalyst 15.0, tvOS 15.0, watchOS 8.0, *)
-extension SafetySetting.BlockThreshold: Codable {
-  public init(from decoder: Decoder) throws {
-    let value = try decoder.singleValueContainer().decode(String.self)
-    guard let decodedThreshold = SafetySetting.BlockThreshold(rawValue: value) else {
-      Logging.default
-        .error("[FirebaseVertexAI] Unrecognized BlockThreshold with value \"\(value)\".")
-      self = .unknown
-      return
-    }
-
-    self = decodedThreshold
-  }
-}
-
-@available(iOS 15.0, macOS 11.0, macCatalyst 15.0, tvOS 15.0, watchOS 8.0, *)
-extension SafetySetting: Codable {}
+extension SafetySetting: Encodable {}